backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.59.3
python-multipart==0.0.6
json_repair==0.28.2
# PRISM-specific dependencies
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
scipy==1.15.1
faiss-cpu==1.10.0
requests==2.32.3
tiktoken==0.8.0
tenacity==9.0.0 